agent-webmaster-py-old/requirements.txt
azure-ai-projects
azure-ai-agents
azure-identity
azure-storage-blob
python-dotenv